Faster key-point matching and speed extraction class
We move speed extraction to its own class SpeedExtractor which we instantiate with every instance of SpeedTest

We make point matching a space intensive O(N) operation instead of the O(N^4) operation that it was before. It still isn't that fast, we doa lot of book-keeping and object creation to make it happen.

TODO: We made it faster so that we could get the speed across more kee point matches. However, due to the texture or _something_, we have quite a few outliers which threaten any basic speed extraction (mean is heavily skewed, as is median to an extent). We might need to do some sort of RANSAC-like operation.